+++ game/game/playercontrol.cpp
@@ -1,5 +1,5 @@
 #include "playercontrol.h"
-
+#include "gamepad.h"
 #include <cmath>
 #include <SDL2/SDL.h>
 #include <iostream>
@@ -1110,185 +1110,3 @@ void PlayerControl::processAutoRotate(Npc& pl, float& rot, uint64_t dt) {
       }
     }
   }
-
-
-void PlayerControl::handleControllerInput() {
-    static bool controllerDetected = false;  // Static flag to track if controller is already detected
-
-    static bool menuActive = false; // Track whether the radial menu is active
-    static int selectedOption = 0;  // Index of the selected menu op
-  
-    // Detect the controller only once
-    if (SDL_NumJoysticks() < 1) {
-        std::cerr << "No joystick or controller detected!" << std::endl;
-        return;
-    }
-
-    // Only print the message the first time the controller is detected
-    if (!controllerDetected) {
-        std::cout << "Controller detected: " << SDL_JoystickNameForIndex(0) << std::endl;
-        controllerDetected = true;
-    }
-
-    // Attempt to load the controller mappings
-    if (SDL_GameControllerAddMappingsFromFile("gamecontrollerdb.txt") < 0) {
-        std::cerr << "Failed to load controller mappings: " << SDL_GetError() << std::endl;
-    }
-
-    // Open the first controller
-    SDL_GameController* controller = SDL_GameControllerOpen(0);
-    if (controller == nullptr) {
-        std::cerr << "Unable to open controller: " << SDL_GetError() << std::endl;
-        return;
-    }
-    
-    const int DEADZONE = 8000;  // Deadzone for analog sticks
-
-    // Get the axis values (left analog stick)
-    int leftX = SDL_GameControllerGetAxis(controller, SDL_CONTROLLER_AXIS_LEFTX);
-    int leftY = SDL_GameControllerGetAxis(controller, SDL_CONTROLLER_AXIS_LEFTY);
-
-    // Process movement if they exceed the deadzone
-    if (abs(leftX) > DEADZONE) {
-        if (leftX > 0) {
-            movement.strafeRightLeft.main[0] = true;
-        } else if (leftX < 0) {
-            movement.strafeRightLeft.reverse[0] = true;
-        }
-    } else {
-        movement.strafeRightLeft.reset();
-    }
-
-    if (abs(leftY) > DEADZONE) {
-        if (leftY > 0) {
-            movement.forwardBackward.reverse[0] = true;
-        } else if (leftY < 0) {
-            movement.forwardBackward.main[0] = true;
-        }
-    } else {
-        movement.forwardBackward.reset();
-    }
-  
-    // Check for D-pad Up (Weapon Melee Action)
-    if (SDL_GameControllerGetButton(controller, SDL_CONTROLLER_BUTTON_DPAD_UP)) {
-        onKeyPressed(Action::WeaponMele, Tempest::KeyEvent::K_Space, KeyCodec::Mapping::Primary);
-    }
-
-    // Check for D-pad Right (Weapon Bow Action)
-    if (SDL_GameControllerGetButton(controller, SDL_CONTROLLER_BUTTON_DPAD_RIGHT)) {
-        onKeyPressed(Action::WeaponBow, Tempest::KeyEvent::K_Space, KeyCodec::Mapping::Primary);
-    }
-
-    // Check for D-pad Down (Weapon Magic Action)
-    if (SDL_GameControllerGetButton(controller, SDL_CONTROLLER_BUTTON_DPAD_DOWN)) {
-        // For Magic, cycle through available spells (WeaponMage3 to WeaponMage10)
-        static int currentMagicSlot = Action::WeaponMage3;
-        onKeyPressed(static_cast<Action>(currentMagicSlot), Tempest::KeyEvent::K_Space, KeyCodec::Mapping::Primary);
-        // Cycle through magic slots
-        currentMagicSlot++;
-        if (currentMagicSlot > Action::WeaponMage10) {
-            currentMagicSlot = Action::WeaponMage3;
-        }
-    }
-
-    if (SDL_GameControllerGetButton(controller, SDL_CONTROLLER_BUTTON_A)) {
-        // A button pressed (map to Weapon action)
-        onKeyPressed(KeyCodec::Action::Weapon, Tempest::KeyEvent::KeyType::K_Return, KeyCodec::Mapping());
-    }
-    if (SDL_GameControllerGetButton(controller, SDL_CONTROLLER_BUTTON_B)) {
-        // If B button is pressed, trigger Jump action
-        ctrl[Action::Jump] = true;  // Assuming ctrl is an array of actions
-    } else {
-        ctrl[Action::Jump] = false; // Reset the jump action when B button is not pressed
-    }
-
-  // Check if Left Stick (L3) is pressed to activate the menu
-    if (SDL_GameControllerGetButton(controller, SDL_CONTROLLER_BUTTON_LEFTSTICK)) {
-        menuActive = true;
-    } else {
-        menuActive = false;
-    }
-
-    if (menuActive) {
-        // Menu is active, let's display the radial menu and handle navigation
-
-        // Get the right joystick values (right analog stick)
-        int rightX = SDL_GameControllerGetAxis(controller, SDL_CONTROLLER_AXIS_RIGHTX);
-        int rightY = SDL_GameControllerGetAxis(controller, SDL_CONTROLLER_AXIS_RIGHTY);
-
-        // Deadzone for the right joystick to prevent unintentional movement
-        const int DEADZONE = 8000;
-
-        // Only process the right joystick movement if it exceeds the deadzone
-        if (abs(rightX) > DEADZONE || abs(rightY) > DEADZONE) {
-    // Calculate the angle of the right joystick movement
-    float angle = atan2f(static_cast<float>(rightY), static_cast<float>(rightX)); // Get the angle in radians
-
-    // Normalize the angle to degrees (0° to 360°)
-    angle = angle * 180.0f / static_cast<float>(M_PI);
-
-    // Ensure the angle is between 0° and 360°
-    if (angle < 0) {
-        angle += 360.0f;  // Adjust negative angles to positive range
-    }
-
-    // Determine the selected option based on the angle
-    if (angle >= 0 && angle < 45) {
-        selectedOption = 0; // Option 1 (right)
-    } else if (angle >= 45 && angle < 135) {
-        selectedOption = 1; // Option 2 (down)
-    } else if (angle >= 135 && angle < 225) {
-        selectedOption = 2; // Option 3 (left)
-    } else if (angle >= 225 && angle < 315) {
-        selectedOption = 3; // Option 4 (up)
-    } else {
-        selectedOption = 0; // Wrap around to Option 1 (right)
-    }
-
-    // Optionally, print or use the angle
-    std::cout << "Joystick angle: " << angle << "°" << std::endl;
-}
-
-            // Optional: Visualize the selected option in the console (or UI)
-            std::cout << "Selected Option: " << selectedOption << std::endl;
-        }
-
-        // Check if the user selects an option (e.g., pressing the A button)
-        if (SDL_GameControllerGetButton(controller, SDL_CONTROLLER_BUTTON_A)) {
-            // Execute the action for the selected option
-            switch (selectedOption) {
-                case 0:
-                    std::cout << "Option 1 selected" << std::endl;
-                    // Trigger corresponding action for option 1
-                    break;
-                case 1:
-                    std::cout << "Option 2 selected" << std::endl;
-                    // Trigger corresponding action for option 2
-                    break;
-                case 2:
-                    std::cout << "Option 3 selected" << std::endl;
-                    // Trigger corresponding action for option 3
-                    break;
-                case 3:
-                    std::cout << "Option 4 selected" << std::endl;
-                    // Trigger corresponding action for option 4
-                    break;
-            }
-        }
-    
-
-    if (SDL_GameControllerGetButton(controller, SDL_CONTROLLER_BUTTON_LEFTSHOULDER)) {
-        movement.strafeRightLeft.reverse[0] = true;
-    } else {
-        movement.strafeRightLeft.reverse[0] = false;
-    }
-
-    // Check if the right shoulder button is pressed
-    if (SDL_GameControllerGetButton(controller, SDL_CONTROLLER_BUTTON_RIGHTSHOULDER)) {
-        movement.strafeRightLeft.main[0] = true;
-    } else {
-        movement.strafeRightLeft.main[0] = false;
-    }
-  
-    SDL_GameControllerClose(controller);  // Close the controller
-}
